Move JUnit-related code into new `artifact` package
